admin_app/requirements.txt
psycopg2==2.9.3
fastapi==0.81.0
sqladmin==0.5.0
loguru==0.6.0
pydantic==1.9.1
python-dotenv==0.20.0
SQLAlchemy==1.4.38
uvicorn==0.18.3
PyJWT==1.7.1
